+++ common/src/main/java/immersive_aircraft/Entities.java
@@ -15,36 +15,42 @@ public interface Entities {
     Supplier<EntityType<GyrodyneEntity>> GYRODYNE = register("gyrodyne", EntityType.Builder
             .of(GyrodyneEntity::new, MobCategory.MISC)
             .sized(1.25f, 0.6f)
+            .clientTrackingRange(12)
             .fireImmune()
     );
 
     Supplier<EntityType<BiplaneEntity>> BIPLANE = register("biplane", EntityType.Builder
             .of(BiplaneEntity::new, MobCategory.MISC)
             .sized(1.75f, 0.85f)
+            .clientTrackingRange(12)
             .fireImmune()
     );
 
     Supplier<EntityType<AirshipEntity>> AIRSHIP = register("airship", EntityType.Builder
             .of(AirshipEntity::new, MobCategory.MISC)
             .sized(1.5f, 2.5f)
+            .clientTrackingRange(12)
             .fireImmune()
     );
 
     Supplier<EntityType<CargoAirshipEntity>> CARGO_AIRSHIP = register("cargo_airship", EntityType.Builder
             .of(CargoAirshipEntity::new, MobCategory.MISC)
             .sized(1.75f, 2.5f)
+            .clientTrackingRange(12)
             .fireImmune()
     );
 
     Supplier<EntityType<QuadrocopterEntity>> QUADROCOPTER = register("quadrocopter", EntityType.Builder
             .of(QuadrocopterEntity::new, MobCategory.MISC)
             .sized(1.5f, 0.5f)
+            .clientTrackingRange(10)
             .fireImmune()
     );
 
     Supplier<EntityType<BlimpEntity>> BLIMP = register("blimp", EntityType.Builder
             .of(BlimpEntity::new, MobCategory.MISC)
             .sized(8.0f, 8.0f)
+            .clientTrackingRange(12)
             .fireImmune()
     );